activate commented out tests again and removed sqlite fix
The round was removed because the new calculation method should
not need to be rounded. Unfortunately this only works with
databases which enforce a decimal setting (2 digits) and not
with sqlite. Also postgres is more picky and we have to use
strings if the datatype is a string to search for it.